Fix more bugs.Added more features!
* Added the feature of Pig
* Added the feature of Sheep
* Sheep can eat grass
* You can shear wool
* Sheep have more colors

+++ src/main/java/me/onebone/actaeon/hook/AnimalGrowHook.java
@@ -17,7 +17,7 @@ import me.onebone.actaeon.entity.MovingEntity;
  * ||||
  */
 public class AnimalGrowHook extends MovingEntityHook {
-    private int growTick;
+    private long growTick;
 
     public AnimalGrowHook(MovingEntity animal, int growTick) {
         super(animal);
